Update 2: Detection Algorithms Added
I added three pop-up messages that tell the user if someone won a TicTacToe, or if there is a tie. Also fixed the obvious bug that was X has the advantage of getting the click first. I later changed the double-click option for the button to a popup dialogue that asks the user what mark should there be.